+++ packages/react/src/components/SideBar/index.tsx
@@ -1,77 +1,22 @@
-import React, { ReactElement, type ReactNode } from 'react';
+import React, { useState, ReactNode } from 'react';
 import clsx from 'clsx';
 import { Text } from '../Typography/Text';
+import { Icon } from '../Icons';
+import { NavItem, NavSubItem } from './SubNav/NavLink';
+import { SubNav } from './SubNav';
 
-interface NavSubItem {
-  label: string;
-  path: string;
-  icon: ReactElement;
-  notification?: number;
-}
-
-interface NavItem {
-  title: string;
-  subItems: NavSubItem[];
-}
-interface SubNav extends NavItem {
-  activeItem: string;
-}
 interface SideBarProps {
-  logo: ReactNode;
   activeItem: string;
   items: NavItem[];
+  logo?: ReactNode;
   footer?: ReactNode;
-  width?: string;
+  width?: string | number;
+  padding?: string | number;
+  buttonsRadius?: string;
+  collapseText?: string;
+  isCollapsed?: boolean;
+  onToggleCollapse?: (collapsed: boolean) => void;
 }
-
-const SubNav = ({ title, subItems, activeItem }: SubNav) => (
-  <div className="flex flex-col gap-y-2">
-    <Text size="xs" color="text-neutral-600" weight="regular">
-      {title}
-    </Text>
-    {subItems.map((subItem, index) => (
-      <NavLink key={index} subItem={subItem} activeItem={activeItem} />
-    ))}
-  </div>
-);
-
-const NavLink = ({
-  subItem,
-  activeItem,
-}: {
-  subItem: NavSubItem;
-  activeItem: string;
-}) => {
-  const isActive = activeItem === subItem.path;
-
-  const linkClasses = clsx(
-    'flex w-full flex-row items-center gap-x-2 p-2 rounded-[5px]',
-    'transition-colors duration-150 ease-in-out',
-    'hover:text-white hover:primary-bg',
-    {
-      'bg-transparent text-neutral-800': !isActive,
-      'primary-bg text-white': isActive,
-    },
-  );
-
-  return (
-    <a className={linkClasses} href={subItem.path}>
-      <div className="flex w-full flex-row items-center gap-x-2">
-        {React.cloneElement(subItem.icon, { color: 'inherit' })}
-        <Text size="sm" color="inherit" weight="regular">
-          {subItem.label}
-        </Text>
-      </div>
-      {!!subItem.notification && (
-        <div className="rounded-full bg-error flex items-center justify-center px-[5px]">
-          <Text size="xs" color="text-white">
-            {subItem.notification}
-          </Text>
-        </div>
-      )}
-    </a>
-  );
-};
 
 const SideBar = ({
   logo,
@@ -79,32 +24,100 @@ const SideBar = ({
   footer,
   activeItem,
   width = '250px',
+  padding = '4px',
+  buttonsRadius = '5px',
+  isCollapsed: controlledCollapsed,
+  collapseText,
+  onToggleCollapse,
 }: SideBarProps): JSX.Element => {
-  const sidebarClasses = clsx(
-    'p-4 h-screen bg-white flex flex-col justify-between',
-  );
+  const [internalCollapsed, setInternalCollapsed] = useState(false);
+  const isControlled = controlledCollapsed !== undefined;
+  const isCollapsed = isControlled ? controlledCollapsed : internalCollapsed;
+
+  const toggleSidebar = () => {
+    if (isControlled && onToggleCollapse) {
+      onToggleCollapse(!controlledCollapsed);
+    } else {
+      setInternalCollapsed(!internalCollapsed);
+    }
+  };
 
   return (
     <div
-      className={sidebarClasses}
-      style={{
-        width: width,
-      }}
+      style={
+        isCollapsed
+          ? {
+              padding,
+              width: 60,
+            }
+          : {
+              padding,
+              width,
+            }
+      }
+      className={clsx(
+        'h-full bg-white flex flex-col justify-between  transition-all delay-150 duration-300',
+      )}
     >
       <div>
-        <div className="flex justify-center items-center border-b-[0.5px] pb-4 mb-4">
-          {logo}
-        </div>
+        {logo && (
+          <div className="flex justify-center items-center border-b-[0.5px] py-4 mb-4">
+            {logo}
+          </div>
+        )}
         <nav className="flex flex-col gap-y-2">
           {items.map((item, index) => (
-            <SubNav key={index} {...item} activeItem={activeItem} />
+            <SubNav
+              radius={buttonsRadius}
+              key={index}
+              {...item}
+              activeItem={activeItem}
+              isCollapsed={isCollapsed}
+            />
           ))}
         </nav>
       </div>
-      <footer>{footer}</footer>
+      <footer>
+        {!isCollapsed && footer}
+        <div className="flex justify-center items-center border-t-[0.5px] mt-4">
+          <button
+            onClick={toggleSidebar}
+            className="p-2 bg-transparent w-full "
+            style={{
+              borderRadius: buttonsRadius,
+            }}
+          >
+            {isCollapsed ? (
+              <div className="flex justify-center items-center gap-4 flex-row text-inherit">
+                <Icon name="arrow-right-double" size={20} color="inherit" />
+              </div>
+            ) : (
+              <div className="flex justify-center items-center gap-4 flex-row text-inherit">
+                <Icon name="arrow-left-double" size={20} color="inherit" />
+                {collapseText && (
+                  <Text
+                    size="sm"
+                    color="inherit"
+                    weight="medium"
+                    style={{
+                      textAlign: 'center',
+                      display: 'flex',
+                      flexWrap: 'wrap',
+                      whiteSpace: 'nowrap',
+                    }}
+                  >
+                    {collapseText}
+                  </Text>
+                )}
+              </div>
+            )}
+          </button>
+        </div>
+      </footer>
     </div>
   );
 };
+
 SideBar.displayName = 'SideBar';
 
 export { SideBar, SideBarProps, NavItem, NavSubItem };